requirements.txt
python-redmine==2.3.0
requests==2.26.0
python-dotenv==0.19.0
azure-devops==7.1.0b3